+++ joyreactor-download-button/CHANGELOG.md
@@ -1,16 +1,26 @@
 # JoyReactor Download Button Changelog
 
+<!--
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+### BREAKING CHANGES
+-->
+
 ## 1.2.0 - 2018-01-16
 
 ### Changed
 
-- Refactoring
+- [Internal] Refactoring
 
 ## 1.1.0 - 2017-11-26
 
 ### Changed
 
-- Build with rollup
+- [Internal] Build with rollup
 
 ## 1.0.0 - 2017-11-17