+++ pornolab-english/CHANGELOG.md
@@ -1,6 +1,18 @@
 # English Pornolab Changelog
 
+<!--
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+### BREAKING CHANGES
+-->
+
 ## 0.1.0 - YYYY-MM-DD
+
 ## Added
-- Base functionlity
+
+- Base functionality
 - Topic page